vehicle-insurance-agentic-ai/requirements.txt
streamlit
boto3
pandas
snowflake-snowpark-python
langchain
langchain-community
langgraph